Split usage module into two, with one module containing the core code and another the code for rendering the questions and outputs.
